# src/main/java/com/fulfilment/application/monolith/stores/adapters/restapi/dto/PatchStoreRequest.java
package com.fulfilment.application.monolith.stores.adapters.restapi.dto;

import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.Pattern;

/**
 * Request DTO for partially updating a store (PATCH operation).
 */

public record PatchStoreRequest(
    @Pattern(regexp = ".*\\S.*", message = "Store name must not be blank.")
    String name,
    @Min(value = 0, message = "Store quantityProductsInStock cannot be negative.")
    Integer quantityProductsInStock
) {}